move WebbColorsType to webb-ui-components
