+++ packages/page-statistics/types/index.d.ts
@@ -52,15 +52,3 @@ export interface SocialConfigsType extends Omit<ExternalLink, 'name'> {
    */
   Icon: (props: IconBase) => JSX.Element;
 }
-
-/**
- * The Webb color type
- */
-export interface WebbColorsType {
-  mono: Record<string | number, string>;
-  purple: Record<string | number, string>;
-  blue: Record<string | number, string>;
-  green: Record<string | number, string>;
-  yellow: Record<string | number, string>;
-  red: Record<string | number, string>;
-}
