+++ packages/webb-ui-components/src/types/index.d.ts
@@ -28,7 +28,6 @@ declare module '@tanstack/table-core' {
 /**
  * The base interface required all component to extends in their props
  */
-
 export interface WebbComponentBase extends React.HTMLAttributes<HTMLElement> {
   /**
    * The tailwindcss className to override the style
@@ -42,6 +41,18 @@ export interface WebbComponentBase extends React.HTMLAttributes<HTMLElement> {
    * Children node
    */
   children?: React.ReactNode;
+}
+
+/**
+ * The Webb color type
+ */
+export interface WebbColorsType {
+  mono: Record<string | number, string>;
+  purple: Record<string | number, string>;
+  blue: Record<string | number, string>;
+  green: Record<string | number, string>;
+  yellow: Record<string | number, string>;
+  red: Record<string | number, string>;
 }
 
 /**
